fix(parser) : add `ERROR` to endwordlist before valid check
1) Problem

If there are parent node which is ERROR, plugins misunderstands it is
endwised already.

2) Why

When `get_root_node()` executes, it checks nodes only in endwordlist,
If there are ERROR, it doesn't think it is parent node.

3) Solution

Add `ERROR` node before valid check, It can be used in get_root and
check endwised already.